ml/pyproject.toml
[project]
name = "engin33ring-thesis-ml"
version = "0.1.0"
description = "Road sign detection model training pipeline"
requires-python = ">=3.11"
dependencies = [
    "ultralytics>=8.3.0",
    "torch>=2.2.0",
    "torchvision>=0.17.0",
    "opencv-python>=4.9.0",
    "albumentations>=1.4.0",
    "matplotlib>=3.8.0",
    "pandas>=2.2.0",
    "pyyaml>=6.0.0",
    "onnx>=1.16.0",
    "onnxruntime>=1.17.0",
    "kagglehub>=0.3.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
]

[tool.ruff]
line-length = 88
indent-width = 4

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

